+++ assignment1/src/FEManagementHandler.java
@@ -22,7 +22,7 @@ public class FEManagementHandler implements FEManagement.Iface {
 
     public PerfCounters getPerfCounters() {
 
-        perfCounter.numSecondsUp = System.currentTimeMillis() - serviceUpTime;
+        perfCounter.numSecondsUp = (int)(System.currentTimeMillis() - serviceUpTime);
         perfCounter.numRequestsReceived = perfManager.numRequestsReceived;
         perfCounter.numRequestsCompleted = perfManager.numRequestsCompleted;
